Create slide module - Add __init__.py, mybin.py executable - Rewrite imports - Split join o join_core and join_utils - Use module docstrings instead of comments
