+++ mybin.py
@@ -1,0 +1,6 @@
+#/usr/bin/python3
+
+import sys
+from slide.entailment import main
+
+main(sys.argv)
