+++ slide-wrapper-orig.sh
