+++ slide/__init__.py
@@ -1,0 +1,5 @@
+"""SLIDE is a prototype tool for checking entailment in Separation Logic with
+user-provided inductive definitions of recursive data structures (lists, trees,
+and beyond)
+Home Page: http://www.fit.vutbr.cz/research/groups/verifit/tools/slide/
+"""
